Able to add multiple ships of different sizes to each board.
The placeShip* functions appear to need to be added before the script starts.
Buttons for all possible ship sizes, 1-6 will need to be present when the game starts.
If you want to make it clear to the user not to add those ships outside their range,
neuter those buttons somehow (maybe add some dummy function to it? idk)
Give visual clue.

If buttons are addeed dynamically, the eventListeners are added all at once to the cells, and removed all at once.
This places every ship in the same location (promise workaround?)